REFACTORING: Тип id заменён с Int на Long

+++ data/appData/api/src/main/java/kanti/todoer/data/appdata/AppDataLocalDataSource.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface AppDataLocalDataSource {
 
-	val currentPlanId: Flow<Int?>
+	val currentPlanId: Flow<Long?>
 
-	suspend fun setCurrentPlan(planId: Int?)
+	suspend fun setCurrentPlan(planId: Long?)
 }
